// constants/enum.ts
export enum E_LABEL {
  TEXT_TITLE = 'TEXT_TITLE',
  BULLETED_LIST_CIRCLE_LEVEL_ONE = 'BULLETED_LIST_CIRCLE_LEVEL_ONE',
  BULLETED_LIST_CIRCLE_LEVEL_TWO = 'BULLETED_LIST_CIRCLE_LEVEL_TWO',
  HIGH_LIGHT = 'HIGH_LIGHT',
  HIGH_LIGHT_ORANGE_RED = 'HIGH_LIGHT_ORANGE_RED',
  HIGH_LIGHT_ORANGE_YELLOW = 'HIGH_LIGHT_ORANGE_YELLOW',
  HIGH_LIGHT_TEAL_GREEN = 'HIGH_LIGHT_TEAL_GREEN',
  TEXT_BANNER_LARGE = 'TEXT_BANNER_LARGE',
  SUBHEADING = 'SUBHEADING',
  LARGE_TITLE = 'LARGE_TITLE',
  CASE_STUDY_TITLE = 'CASE_STUDY_TITLE',
  CASE_STUDY_SUB_TITLE = 'CASE_STUDY_SUB_TITLE',
  CASE_STUDY_DESCRIPTION = 'CASE_STUDY_DESCRIPTION'
}

export enum E_SCREEN_SIZE {
  SM = 640,
  MD = 768,
  LG = 1024,
  XL = 1280,
  TWO_XL = 1536
}

export enum eQueryKey {
  SEARCH = 'search',
  PAGE = 'page'
}